updating

+++ stayhub/src/main/java/com/example/rani/stayhub/service/RoomServiceImpl.java
@@ -4,12 +4,15 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.example.rani.stayhub.dto.RoomDto;
 import com.example.rani.stayhub.entity.Hotel;
 import com.example.rani.stayhub.entity.Room;
+import com.example.rani.stayhub.entity.User;
 import com.example.rani.stayhub.exception.ResourceNotFoundException;
+import com.example.rani.stayhub.exception.UnAuthorisedException;
 import com.example.rani.stayhub.repository.HotelRepository;
 import com.example.rani.stayhub.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
@@ -31,6 +34,11 @@ public class RoomServiceImpl implements RoomService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " +hotelId));
+                 User user = (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id"+ hotelId);
+        }
         Room room = modelMapper.map(roomDto , Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -48,6 +56,10 @@ public class RoomServiceImpl implements RoomService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " +hotelId));
+        User user = (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id"+ hotelId);
+        }
       return hotel.getRooms()
             .stream()
             .map((element)-> modelMapper.map(element, RoomDto.class))
@@ -69,7 +81,11 @@ public class RoomServiceImpl implements RoomService {
        Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: " +roomId));
-                
+
+        User user = (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id"+ roomId);
+        }       
        inventoryService.deleteAllInventories(room);
        roomRepository.deleteById(roomId);
     }
